+++ mitra_activitypub/src/contexts.rs
@@ -32,8 +32,6 @@ pub fn build_default_context() -> Context {
             ("sensitive", "as:sensitive"),
             ("toot", MASTODON_CONTEXT),
             ("Emoji", "toot:Emoji"),
-            ("litepub", "http://litepub.social/ns#"),
-            ("EmojiReact", "litepub:EmojiReact"),
         ]),
     )
 }