+++ mitra_adapters/src/init.rs
@@ -1,3 +1,6 @@
+use std::os::unix::fs::MetadataExt;
+use std::path::Path;
+
 use apx_core::{
     crypto::{
         eddsa::{
@@ -62,6 +65,40 @@ pub fn initialize_app(
         log::warn!("{}", warning);
     };
     config
+}
+
+extern "C" {
+    fn geteuid() -> u32;
+}
+
+fn check_directory_owner(path: &Path) -> () {
+    let metadata = std::fs::metadata(path)
+        .expect("can't read file metadata");
+    let owner_uid = metadata.uid();
+    let current_uid = unsafe { geteuid() };
+    if owner_uid != current_uid {
+        panic!(
+            "{} owner ({}) is different from the current user ({})",
+            path.display(),
+            owner_uid,
+            current_uid,
+        );
+    };
+}
+
+pub fn check_app_directories(config: &Config) -> () {
+    if !config.storage_dir.exists() {
+        panic!("storage directory does not exist");
+    };
+    check_directory_owner(&config.storage_dir);
+    if let Some(ref web_client_dir) = config.web_client_dir {
+        if !web_client_dir.exists() {
+            panic!(
+                "web client directory does not exist: {}",
+                web_client_dir.display(),
+            );
+        };
+    };
 }
 
 // Panics on errors